Refactor regex parser
